+++ apps/frontend/src/app/5_pages/RewardsPage/components/Staking/components/WithdrawLiquidFee/WithdrawLiquidFee.tsx
@@ -3,9 +3,10 @@ import React, { FC, useCallback, useMemo } from 'react';
 import { t } from 'i18next';
 
 import { SupportedTokens } from '@sovryn/contracts';
-import { Button, ButtonType, ButtonStyle } from '@sovryn/ui';
+import { Button, ButtonType, ButtonStyle, Tooltip } from '@sovryn/ui';
 
 import { TransactionType } from '../../../../../../3_organisms/TransactionStepDialog/TransactionStepDialog.types';
+import { MS } from '../../../../../../../constants/general';
 import { useTransactionContext } from '../../../../../../../contexts/TransactionContext';
 import { useAssetBalance } from '../../../../../../../hooks/useAssetBalance';
 import { useGetProtocolContract } from '../../../../../../../hooks/useGetContract';
@@ -13,6 +14,9 @@ import { useMaintenance } from '../../../../../../../hooks/useMaintenance';
 import { translations } from '../../../../../../../locales/i18n';
 import { getRskChainId } from '../../../../../../../utils/chain';
 import { decimalic } from '../../../../../../../utils/math';
+import { MAX_LIQUID_STAKES } from '../../Staking.constants';
+import useGetFilteredDates from '../../hooks/useGetFilteredDates';
+import useGetVestingAddresses from '../../hooks/useGetVestingAddresses';
 
 type WithdrawLiquidFeeProps = {
   amountToClaim: string;
@@ -25,6 +29,24 @@ export const WithdrawLiquidFee: FC<WithdrawLiquidFeeProps> = ({
   lastWithdrawalInterval,
   refetch,
 }) => {
+  const currentDate = useMemo(() => Math.ceil(new Date().getTime() / MS), []);
+  const vestingAddresses = useGetVestingAddresses();
+  const filteredDates = useGetFilteredDates(vestingAddresses);
+
+  const datesLessThanCurrentTime = useMemo(
+    () =>
+      filteredDates.filter(
+        dateInSeconds => parseInt(dateInSeconds) < currentDate,
+      ),
+    [filteredDates, currentDate],
+  );
+
+  const isAboveThreshold = useMemo(
+    () =>
+      datesLessThanCurrentTime.length > MAX_LIQUID_STAKES &&
+      vestingAddresses.length > 0,
+    [datesLessThanCurrentTime, vestingAddresses],
+  );
   const { setTransactions, setIsOpen, setTitle } = useTransactionContext();
 
   const { checkMaintenance, States } = useMaintenance();
@@ -45,11 +67,18 @@ export const WithdrawLiquidFee: FC<WithdrawLiquidFeeProps> = ({
 
   const isClaimDisabled = useMemo(
     () =>
+      isAboveThreshold ||
       claimLiquidSovLocked ||
       !hasTokenBalance ||
       rewardsLocked ||
       decimalic(amountToClaim).isZero(),
-    [amountToClaim, claimLiquidSovLocked, hasTokenBalance, rewardsLocked],
+    [
+      amountToClaim,
+      claimLiquidSovLocked,
+      hasTokenBalance,
+      rewardsLocked,
+      isAboveThreshold,
+    ],
   );
 
   const onComplete = useCallback(() => {
@@ -88,14 +117,26 @@ export const WithdrawLiquidFee: FC<WithdrawLiquidFeeProps> = ({
   ]);
 
   return (
-    <Button
-      type={ButtonType.button}
-      style={ButtonStyle.secondary}
-      text={t(translations.rewardPage.stabilityPool.actions.withdraw)}
-      onClick={onSubmit}
-      disabled={isClaimDisabled}
-      className="w-full lg:w-auto"
-      dataAttribute="liquid-button"
+    <Tooltip
+      children={
+        <div>
+          <Button
+            type={ButtonType.button}
+            style={ButtonStyle.secondary}
+            text={t(translations.rewardPage.stabilityPool.actions.withdraw)}
+            onClick={onSubmit}
+            disabled={isClaimDisabled}
+            className="w-full lg:w-auto"
+            dataAttribute="liquid-button"
+          />
+        </div>
+      }
+      content={t(translations.rewardPage.staking.liquidityMiningError, {
+        count: MAX_LIQUID_STAKES,
+      })}
+      dataAttribute="liquid-button-tooltip"
+      className="inline-block"
+      disabled={!isAboveThreshold}
     />
   );
 };